+++ lib/pragma/association_includer.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Pragma
-  module AssociationIncluder
-    include Adaptor::Loader
-    register ActiveRecord, Poro
-  end
-end