+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Removed `Pragma::Decorator::Error`, which was moved to
   [pragma-decorator](https://github.com/pragmarb/pragma-decorator)
+- Removed `Pragma::AssociationIncluder`, which was moved to
+  [pragma-decorator](https://github.com/pragmarb/pragma-decorator)
 
 ## [2.5.0]